Use goreleaser to cleanup the release building
Without this, some parameters need to be passed around
the different tooling at different point of the lifecycle
(e.g. during the building of the containers).
The current code is therefore stuck to amd64.

This is a problem, as it makes it far less efficient to
run on macos arm processors (we have to use a vm, cannot
run locally).

This fixes it by making sure the images will be able
to be built on all os and architectures by using
the 'goreleaser' abstracting tool. This tool
was introduced into the repository and maintained
but not configured in 2021 [1].

By configuring goreleaser, we can pick architectures
(arm64 and amd64_v1 chosen), and we define ldflags+timestamp
for generating reproducible builds.

With the name of the binaries being generated, it
then made sense to use a single dockerfile to build all
those images, instead of keeping the 3 dockerfiles.
However, this might impact readability, so I am open
to revert this sub-part of the commit.

At the same time, it fixes an issue of the tests NEVER
running in CI, by ensuring the make test is ran before
even building the image. This should weed out a few
issues and help maintainance.

[1]: https://github.com/ca-gip/kubi/commit/a2bfca6d61cb979a1c162b701c16705a3df157db

Signed-off-by: Jean-Philippe Evrard <open-source@a.spamming.party>