Add expose LDAP groups
Without this patch, we are filtering LDAP groups and take a
decision on what to expose.

This is a problem, as it removes the flexibility of rolebindings
later.

We intend to expose custom role bindings for extra services
(for example Kong), which requires teams to be entering a different
model. In that case, the platform team creates a new cluster role,
but the grantee of the role might come from a deployment tool,
hence outside the operator.

I fixed it by first exposing the groups directly in the token
provider. This means that further commits are required to filter
properly + generate the right token groups directly.