Add basic error handling
This is a first cleanup, tackling the absence of errors.

This is not a problem right now, but will most likely fail
linting in the future.

This fixes it by simply adding basic error != nil pattern,
and does NOT rethink the whole error handling.

It is therefore a temporary workaround to cleanup the code.
The whole error handling needs better treatment later, for
example:
- Removing the multiple useless wrappers to come back to
  simpler error management
- Rethink the signature of the functions to take errors as
  first citizens
- Cleanup code to better handle failures.

These points are NOT included in this commit, and should
be done in a later commit.