Set goreleaser version
Without this, we use different versions of goreleaser in
CI (github actions) and locally.

This is a problem, as it leads to discrepencies between github
actions and our code: More recent versions of goreleaser
have changed their config file syntax.

This fixes it by pinning the version in github actions.