Use goreleaser to push release
Without this, goreleaser is broken, and does not properly push
neither on PR or on tag. This is due to a premature merge of
a previous MR.

This fixes it by ensuring the workflow works. It might need
a few iterations to test the tag system is working as expected.