Fix goreleaser usage
Something went wrong in the commit updating the version for
goreleaser (04785e18f0ff19a7c44e9e6bb7b8d276ab98ec18).

The Makefile was not appropriately updated.
This is a problem, as the goreleaser is not using the right
version, and could lead to failures.

On top of that, the Makefile did not have all the phony inputs
appropriately adapted and goreleaser did not include the right
name of the image... (all of which was supposedly added in that commit).

This fixes it by cleaning up all the code linked to
goreleaser.