Refactor tokenProvider
Without this, the code was duplicated and the generation of
claims was not very readable. For example, it contained steps
that are part of the issuer initialisation.

This is a problem, as it leads to difficult reviews and
difficulty to iterate on the topic.

This fixes it by creating a new constructor for the token
issuer, which clarifies all what's necessary for it, and fatals
if the requirements are not met. This makes sure the code does
not error forever when the requirements are not met.

On top of that, the signature of the token was separated,
allowing easier testing. The testing has shown a lack of
handling the errors in the JWT signature checks, which should
be fixed in a later commit.