Prevent panic on empty private key
Without this, creating a token provider with no private key
will still try to sign the token.
This should never happen.

However, acting on it, even on tests, means a generic panic,
instead of a just an error. This handles the pointer dereferencing
to ensure the code does not panic, as it should already fatal
on main through the constructor.