Move to use a generic basicAuth handler
Without this, the basic auth is included deep in the call stack.

This is a problem, as it means multiple calls have the information
about passwords, and need to carry useless data around.

This fixes it by ensuring a new middleware for basicAuth was
added, directly connecting to ldap. The user information
is then filtered to keep only username, userDN, and email.
It is then passed in a context for use in the next httpHandlers.

At the same time, it allowed me to see that the GenerateConfig
would not fail if the generation of the token results in an
error. I fixed it by adding the same logic in GenerateConfig
and in GenerateJWT http handlers.