Inline ldap single calls
We use different but similiar constructor methods for the
Has* calls.

This is a problem, as it makes the ldap methods unnecessary
harder to read.

This fixes it by ensuring the code is more readable, and allowed
us optimisations, like regrouping code to connect, query,
and return results with our defaults.