Clarify group membership
Without this, one might expect to have groups[] to be
always populated, and passed to JWTClaims generation.
However, this is not the case, as the groups are empty if
the user does not have direct access rights to the cluster.

This is a problem, as it will prevent further evolution of
the group management.

To fix this, I made sure that _ALL THE GROUPS_, including
the special groups (appops, customerops, cloudops, containerops)
have their groups fetched and generated in the JWT.

I also moved the code from ldap to project.go, as none
of the code was actually relevant from ldap perspective:
All the code was manipulating project objects.

Tests were added to ensure the behaviour was intact.

The code also took the opportunity to remove incorrectly
exported functions back to internal functions (and fixing their
tests).