Move CA to main
Presenting the CA is a very small endpoint, lost in the middle
of the "services" internal package.

This is a problem, as it makes it annoying to find. On top of
that, it needed to be passed global variables instead of having
direct access to config data.

This fixes it by making sure this endpoint (only used once) is
directly readable from the main, as it's a two-liner.