Bring constants closer to their usage
Without this, one might wonder where the constants are used.
One might even think that Service account is used to provision
some parts of kubi, where in fact it is only used for auth.

This is a problem, as it could lead to misunderstandings in the
code. In other words, I believe that moving the constants will
make it more explicit about their scope and maintenance.
Therefore it makes clear that the constants moved here
are ONLY usable for auth.

To avoid issues, this removes the constant from generic "utils"
use.