Improve testability
Without this, much code is relying on global variables,
or to state in the different functions.

This is a problem, as it removes the testability, and code
is hard to refactor.

This fixes it by slowly getting away from spaghetti code,
starting by refactoring the data structures.

This prevents leakage of data. It will also allow us
later to get rid of the global variables.

This makes the code more idiomatic, and split the internal
packages based on what they should provide (and how can they
interface).

Therefore, the code should be more readable in the long run.