Apply DRY on code
A few calls are long duplicates.
Without this patch, the code for netpols, rolebindings, and
projects is duplicated in a few methods.

This is a problem, as it makes the code harder to read, and
harder to maintain.

This fixes it by refactoring the calls, making them more
testable (although without adding tests), and making them
simpler to read.