Expose LDAP Group to rolebinding
Without this, we will rely on our legacy group names.

This is a problem, as it means we do not gain the flexibility
brought by the previous "add group" feature automatically in
each project.

This fixes it by ensuring the rolebinding adds a new line for
the users of the group.