Implement arbitrary groups login
Without this, the groups can only come from the pre-existing bases.
This will be a problem, as certain teams (KONG) need to come from
a completely different part of the org, and will not be allowed to
log in.

This fixes it by adding another base for query, which will be at
higher level than the previous ones.
This has an impact on performance, as LDAP will now query on
broader domain groups. This is compensated by NOT doing multiple
smaller queries later, but keeping all the results and filtering
them in this code.

Hence, a sorting/filtering is included in this code.

The filtering of all those groups is made with a regexp, which
MUST be passed in the config.

A sorting of the existing special groups (project, appops, admins...)
is provided for compatibility reasons. Instead of doing more
queries, they are included based on string comparison of DNs.