Clarify Authorization
After a demo with the team, I realised many people did not
understand the value of the filtering, which was explictly
asking the operator to fill the regexp correctly.

This is a problem, as operators would be adding .* and granting
access to everyone with a valid login.

This changes the implementation to be more surgical:
- The allowList now only includes the group names (CNs) instead
  of the whole DN. This is a risk in case of duplicated, but it
  brings readability.
- The authorize part is now explicit in the token generation.