Revert to use full DN instead of CN for autorization.
I do not fully trust our users to do the right thing and avoid
to have the same group names in different parts of the organisation.

If I rely on CN, it means a duplicate name could grant access
incorrectly. This is a problem, and I do not want this refactor
to be causing a security impact.

Instead of fully rewriting the code OR ensuring that the org
have no duplicate CN names, I make sure the regexp is matching
on the full DN.

I expect this should be cleaned in the future, as Kubernetes itself
does not have this notion.