Support special rolebindings
Without this patch, the rolebindings for appops, customerops,
... are only using the old names, not the LDAP groups.

This is a problem, as it will prevent us to clean up the rest
of the code later.

This fixes it by relying on the config of the group base to
know which group to add in the rolebinding.

In the future, it would be best to template those from a
config map: it would help to not being directly interconnected
to LDAP by having a simple config to template for the cluster.
It would also be simpler to manage with a gitops tool.