Cleanup utils logger and metris
There is no point to have multiple loggers and
wrappers:
- they need us to browse the code to realise
  what they do
- they have 0 value compared to slog
- they encouraged a laziness to not use a good
  structured output.

This fixes it by removing zerolog and simply using
slog. To do that, bubbling up errors made it easier
to improve logging, as we now only handle only
errors in a lot of functions, bubbling up errors
to the functions in which the metrics and logs
are generated.

This made it also clear that the vector for the
metrics were also in a weird location, while
they could be near their usage for ease of
understanding.