Perform several targeted ldap groups search instead of a single complex and possibly lengthy one
Without this change, it is tricky to get all of a user's groups that are relevant to k8s in a sensible manner.

We fix this by being explicit about the ldap parent groups whose child groups are expected to be useful for a k8s cluster.

It is a clearer and more maintenable solution. And since there shouldn't be more than a couple of parents groups it should not cause too much overhead due to a few more ldap searches