Fix ldap search with pagination bug
When making a ldap search, the pagination was managed both by the ldap search library and the calling code.
This lead to a fatal error when the number of results exceeded the page size.
This issue has been fixed by removing the pagination management in the calling code. The pagination is thus managed only by the ldap library